nonprofit_grant_writer_dj/requirements.txt
semantic-kernel>=1.28.1
quart>=0.18.0
chainlit>=0.7.0
python-docx>=0.8.11
azure-identity>=1.12.0
azure-search-documents>=11.4.0
qdrant-client>=1.6.0
requests>=2.28.0
python-dotenv>=0.21.0
beautifulsoup4>=4.12.0 
openai>=1.0.0
pydantic>=2.0.0,<3.0.0
hypercorn>=0.14.0
